split outcome
