+++ src/factor_instances_provider/provider/outcome/mod.rs
@@ -1,0 +1,9 @@
+mod factor_instances_provider_outcome;
+mod factor_instances_provider_outcome_for_factor;
+mod internal_factor_instances_provider_outcome;
+mod internal_factor_instances_provider_outcome_for_factor;
+
+pub use factor_instances_provider_outcome::*;
+pub use factor_instances_provider_outcome_for_factor::*;
+pub use internal_factor_instances_provider_outcome::*;
+pub use internal_factor_instances_provider_outcome_for_factor::*;
